Improve shutdown logic for ACP server
Use weakref.finalize to shut down the ACP server when the Python
interpreter exits. Note that there is no strong guarantee that this
will run, it can fail e.g. if there is a segfault in the interpreter,
or if it is killed via SIGKILL.
However, this is superior to calling 'close' in the server's
__del__, which is not guaranteed to execute even if the interpreter
exits cleanly.