Add POD dataclasses, DB implementation, property wrapper
Add dataclasses which describe the data contained in the Model and
ModelingGroup.

Add a 'DB' class, to wrap instantiation of models.

Add property helpers for accessing object properties via gRPC. The
current update mechanism performs a GET request for each __getattr__,
and a GET + PUT request for every __setattr__.

Add a method for loading ACP models from FE files, and extend the
model properties which are exposed.