Add helper to launch ACP in docker container
Rename '_launcher.py' ro '_server.py', and add a helper to
launch ACP in a (local) Docker container.
Convert the '.check()' and '.close()' methods of the LocalAcpServer
to free functions.
Add a helper to wait for the server to respond, and a RemoteAcpServer
which does not manage launching the server, instead just connecting
(insecurely, for now) to a hostname + port.