Add option to run tests with local executable (#28)
Allow tests to be run with a local executable instead of through
the Docker container.
If the '--server-bin' parameter is passed to the 'pytest' run,
the local executable is used. In that case, '--license-server'
can not be passed; we assume that the local executable has
licensing already set up.

Add pytest to 'pre-commit' extra because mypy needs 'pytest' to be
installed for the pre-commit check of the testing code.