Implement a Mapping interface for collections, by ID
Add a Mapping interface (ID -> object) for collections. The
mapping is 'manually' implemented, since there is no built-in
Protocol that fulfills the needs.
The currently implementation is read-only, mutating methods
are not yet implemented.

Rejected alternative: Building the mapping on top of
'collections.abc.Mapping' or the equivalent 'typing.Mapping'. This
is cumbersome because the auto-generated '.values()', '.items()'
etc. assume that the '__getitem__' call is constant cost. Since
we need to first list the objects to get all IDs in '__getitem__',
this is not true.
By manually implementing the methods, the number of gRPC calls
used can be more easily controlled; at the cost of not having
compatibility with the 'collections.abc.Mapping' interface enforced.