Use locked dependencies for CI, add nightly testing from built wheel (#60)
In the regular CI, install pinned dependencies (from `poetry.lock`) 
instead of open dependencies (from the built wheel). This ensures 
that CI works independently from new releases of dependencies on PyPI.

To check that we are still compatible with dependency releases, 
add a nightly workflow that installs from the wheel.

Since the nightly build tests all Python versions, we reduce the
normal CI build to only test the edges of the compatibility
window (3.7 and 3.10).