Instantiate objects without a server connection (#53)
Move the tree object constructor to a 'from_resource_path'
class method. In the main constructor, instead take the object
properties as keyword arguments. This allows instantiating the
objects without a connection to a server.

NOTE: The '.store()' method which allows to then save such an
object onto a server is only implemented for 'Rosette'. This should be
simpler once the 'Create' interface is harmonized between services;
right now, they don't all accept properties.

Change the tree objects to internally track only a 'Channel' object,
instead of a full reference to a 'ServerProtocol'.

Move the tree object definitions into a separate submodule, and
create an abstract base class to implement common functionality.

Add a 'StubWrapper' protocol, which allows using the CRUD stub
operations without knowing the request / reply types. Add a
simple implementation of the protocol.
The goal is to remove this protocol and use the stubs directly once
the '.proto' files are harmonized / simplified.

Change the internal data representation of the tree objects to
directly use the protobuf message class, instead of a separate
dataclass. The dataclasses are removed.

Use the 'StubWrapper' in the grpc property helpers, and
add options to en-/decode the protobuf object to its exposed
Python representation (e.g. numpy array).

Add a helper function that defines the collection methods
('create_*', and the collection property) on the parent.