Implement '.pop' and '.popitem' for collection (#72)
Implement methods `.pop` and `.popitem` on the collection to 
remove an object from the server (keeping only a local representation).

Also add `.clone()` and `.delete()` on the objects itself.

Compared to the `collections.abc.MutableMapping` interface, the 
`__setitem__`, `update` and `setdefault` methods are still missing. 
However, it iss not yet clear if we need `__setitem__` or how it should 
work if we do.