Switch to sphinx_autodoc_typehints for doc (#93)
See discussion in https://github.com/pyansys/dev-guide/issues/187

While using `sphinx.ext.autodoc.typehints` worked for us, following
the common approach seems superior. 

Remove the `_autosummary` directory during `make(.bat) clean`  in
the documentation. This directory is auto-generated, and not
removing it can cause a rebuild of the documentation to be
incomplete.

Use `functools.wraps` for the mapping create method, to get rid of
doc build warnings.

Loosen and update the dependencies in `pyproject.toml`.