Improve object string representation (#97)
Add `__repr__` and `__str__` implementations for tree objects.

The `__repr__` prints the class name and ID (or name if the object doesn't have an ID).

The `__str__` prints the object's attributes. To do so, we add a `GRPC_PROPERTIES`
class method to the tree objects. This is generated by a class decorator, and the
`grpc_*` helper functions emit a custom `property` subclass to know which class
attributes should be printed.