Drop Python 3.7 compatibility (#122)
Drop compatibility for Python 3.7, since its end-of-life (2023-06-27) will most 
likely be before our initial release , or close enough to not matter.

Add `pyupgrade` to the pre-commit hooks, to modernize the code.

Drop `importlib_metadata` and `typing_extensions` dependencies which are
no longer needed, and simplify the `numpy` dependency.

Increase upper limit Python dependency to `<4`, for consistency with other
PyAnsys libraries. This allows users to install the project on newer Python,
even if it is not explicitly tested / supported.

Adding Python 3.11 to CI/CD is currently blocked by a combination of 
* https://github.com/ansys/ansys-api-acp-private/issues/18 -- we do not
  produce a pre-built version of the API library
* https://github.com/ansys/ansys-tools-protoc-helper/issues/29 -- the version
  of protoc used in the protoc-helper is incompatible with 3.11

Adapt `solve_class40` example to changes in `pydpf-composites`.