Remove docker launcher (#128)
Closes #110 

Removes the `docker` launcher, using `docker-compose` in the tests
instead, with the `upload_file` and `download_file` methods.

To be fixed later: the local server should use a temporary directory,
to have the same behavior as docker-compose. Currently this is
hacked into the `Client` method, instead of properly exposed from
the server.