Expose mesh data, add plotting capability.
Add `elemental_data` and `nodal_data` attributes to tree objects, which fetch the
data from the mesh query API.

The returned objects have a method `to_pyvista`, to convert to a plottable object. Currently,
this requires the `Model.mesh` to be passed in; we may consider how this can be avoided.

Optionally, a `component` (name to be improved) can be passed, to select which data to
add to the PyVista mesh. If a vector component is selected, the data is converted to arrows.

The possible string constants for `component` are exposed in the `ElementalDataType` and
`NodalDataType` enums, which are auto-converted from their protobuf equivalent.

Helper classes and functions for wrapping mesh query data are defined in `_mesh_data.py`:

- a base class `MeshDataBase` which implements `to_pyvista`, as well as the construction
  from a mesh query response
- base classes `ElementalData` and `NodalData`, which are the classes to be used to define
  the mesh data classes for each tree object type
- property helpers `elemental_data_property` and `nodal_data_property`, for defining
  the `elemental_data` and `nodal_data` properties, respectively.

The mesh data itself is exposed in a separate `mesh` on the `Model`. Its class `MeshData`
also implements a `to_pyvista` method.

Add a `__slots__` class attribute in some places where it was missing, to disallow setting
attributes that are not explicitly defined on tree objects.