Jvonrick/production ply object (#244)
Implement the production ply frontend.

In order to support read-only objects I simplified the class Hierarchy for TreeObject. Basically there are now three levels:

GrpcObjectBase: Handles the _GRPC_Properties and its string representation
TreeObjectBase: Handles everything related to storing and instanciating the object.
TreeObject / ReadOnlyTreeObject: Handles the different requests using the stubs.
The reason why I changed this was because I would have to add a ReadOnly variant of the RootGrpcObject, which makes the already complex class hierarchy even more complicated. I did not fully understand why we need the intermediate classes in the hierarchy (such as RootGrpcObject). In my understanding these subtypes are only needed because the are used as types elsewhere (for example in the property helpers). But I think it is easier to adress this with structural subtyping. Therefore I added Editable and Gettable protocols that are implemented by the new tree object classes (ReadOnlyTreeObjects and TreeObject). The downside of the new implementation is that there is a little bit of code repetition for the _get methods. We could add small subjects to which we delegate the different requests to avoid this (Similar to StubStore).

Other changes:

Split the Mapping object into a readable and editable part
Split the TreeObjectTester in a readonly and editable part