Add look-up tables (#258)
Add the `LookUpTable1D`, `LookUpTable1DColumn`, `LookUpTable3D` and
`LookUpTable3DColumn` classes. The two column classes use a common 
base, since they have the same attributes.

The column `value_type` is settable only in the constructor, since the API
only allows changing it when the `data` shape is changed to match, or
the `data` is empty. This would make setting it independently as a property 
challenging / potentially confusing.