Wrap gRPC exceptions in a simplified format
Wrap gRPC RpcError to emit a standard Python exception with a
simplified error message (containing only the first line of the
gRPC error details, and the gRPC error code).
The full exception is still available through the traceback, but
this should improve how the error is typically presented to the
user.