Expose all public objects at the module root (#362)
Expose all public objects at the root `ansys.acp.core` of the module.

Remove imports from "underscored" modules from the tests and examples.

It's expected that some of these will subsequently move down a level 
as part of #361.

Work around a limitation in `sphinx_autodoc_typehints` by adding imports
to `numpy` and `ScalarData` where elemental and nodal data types are
defined.

On Linux, use the `-j auto` sphinx option to build in parallel. On Windows, this
didn't seem to have an effect.