Add options for partial documentation builds (#368)
Add the following options to run a partial doc build, controlled
via environment variables:
- `PYACP_DOC_SKIP_GALLERY=true`: do not build the examples
  gallery
- `PYACP_DOC_SKIP_API=true`: do not build the API reference

Add a quick variant of the documentation build in CI which only
builds the main pages (no API / examples). This build runs in parallel
with the full documentation build, to provider quicker feedback
on errors.

Add docstrings to the enum classes.