Improve error when unavailable material property is accessed
When a material property which is currently unavailable (due to
the difference between isotropic / orthotropic material property
sets) is accessed (getter or setter), raise an AttributeError with
a specific exception message.
To achieve this, the property helpers for material properties get
two new keyword arguments ``available_on_pb_type`` and
``unavailable_msg``. The former determines which protobuf type
needs to be present for the property to be available, and the
latter gives the error message (enhanced automatically with
the attribute name).