Add 'connect' launch mode
Add a launch mode which connects to an existing ACP + Filetransfer
server.
Currently, it is not supported to use only an ACP server (without
filetransfer, e.g. on a local machine).

Adapt the documentation to mention the connect launch mode.

Expose the connect mode in the test configuration via the
--server-urls CLI option.

The immediate use case for the 'connect' launch mode is quick
iterations on tests: connecting to an existing docker-compose
setup avoids the startup time present in both 'direct' and
'docker_compose' mode due to the license checkout.

Other changes:
- In the main README, remove the mention of configuring the launch
  mode, since the default (use unified install) should be enough
  for the common quickstart scenario
- Update poetry lockfile