Improve usability of isotropic / orthotropic material property sets (#414)
Improve the usability of property sets which can be either isotropic
or orthotropic in the following ways (details below)
1. Improve error message on unavailable attributes
2. Allow constructing isotropic property sets directly
3. Add concept documentation

Details:

1.
When a material property which is currently unavailable (due to
the difference between isotropic / orthotropic material property
sets) is accessed (getter or setter), raise an AttributeError with
a specific exception message.
To achieve this, the property helpers for material properties get
two new keyword arguments ``available_on_pb_type`` and
``unavailable_msg``. The former determines which protobuf type
needs to be present for the property to be available, and the
latter gives the error message (enhanced automatically with
the attribute name).

2.
Add classmethod constructors `from_isotropic_constants` and
`from_orthotropic_constants` for directly creating a property
set which is either isotropic or orthotropic. Previously, directly
constructed property sets would always be orthotropic. This
change removes the ability to pass the parameters on the
main constructor (`__init__`).

3.
Add a concept page for material property sets, explaining the
difference between constant / variable property sets, as well
as isotropic / orthotropic. Includes the validation rules for
changing ply type or assigning property sets.

Closes #398