Allow linking to CutoffSelectionRule from ModelingPly
Add the CutoffSelectionRule to the allowed selection rule types
if the parent is not a BooleanSelectionRule.

Fix the docstring of LinkedSelectionRule, since CutoffSelectionRule
and BooleanSelectionRule are disallowed only when the parent is
a BoolenSelectionRule (not for ModelingPly).

Add a check for the operation type when linking to a
CutoffSelectionRule, to avoid a hidden conversion.