Add plotting how-to (#409)
Add a how-to section for plotting the model, showcasing:
- plotting the mesh
- directions plotter
- scalar data plot
- vector data plot (on its own, and with mesh)
- plotting CAD geometries

Add a 'scaling_factor' parameter to 'get_pyvista_glyphs', since this was needed to reasonably show the vector data (ply offset).

Add an example data key for the race car nose `.acph5`.

Since the how-to needs to execute code, it's not run when `PYACP_DOC_SKIP_GALLERY=true` is set. If we need to individually control it, we can always add a separate env var later. In any case, it shouldn't be included in the "quick" CI doc build.

Closes #380