Add conversion from list to numpy array for LookUpTableColumn
Automatically convert lists to numpy arrays for setting lookup
table column data.

Note that using an empty list for directional data is not yet
supported, since we cannot determine the correct shape (0, 3)
in the conversion.
This is intended to be addressed by a backend PR (allowing empty
data with shape (0,) regardless of the target dimensionality),
even though we could *in principle* solve this by injecting
the ``dimension_type`` into the conversion function.

Closes #426.