Revert renaming 'modeling_plies' to 'plies'
Revert the rename of 'modeling_plies' to 'plies'. This change
was originally made to accommodate butt joint sequences and
interface layers in the collection.
However, the current type of the collection (MutableMapping)
is not well-suited to this since it is
- not polymorphic
- not ordered

The current plan is to add MutableMapping exposure also for
``butt_joint_sequences`` and ``interface_layers``, and then
add a _different_ collection as ``plies``, which contains
all different objects which are to be sorted by global ply
number.

This reverts commit 49d68449916c45a39fc6341d602cd3455e050b1a.