Add optimization example
Add an example which optimizes ply angles to minize the maximum
IRF, considering only maximum stress.

The optimization uses the Nelder-Mead (downhill simplex) method
as implemented in scipy.optimize.minimize to find a local minimum.

Since the full optimization takes too long to build as part of
the documentation (~30 min), only a single iteration is performed
as part of the build.