Adapt to changes in material handling (#502)
Adapt to the changes made in material handling:

Material names are deduplicated on creation, and changing a name to an
existing one raises an error. This required changes to the material unittests.

On models which have a link to a MatML file, creating materials is no longer
allowed since storing them back to the MatML file is not implemented (and they
would be lost upon save).
Use a variant of the `minimal_complete_model.acph5` which does not have this
link in tests + doctests to fix this.