Remove upper bounds on dependencies (#531)
Remove upper limits on dependencies. For context, see e.g. the discussion in
https://iscinumpy.dev/post/bound-version-constraints/

In short: downstream projects have no way to remove an artificial (wrong)
upper bound, but adding a missing one is easy. Since it's unknowable
which future versions will be compatible, it's better to err on the side of
allowing them, for the context of libraries.

In addition, we continuously check the newest dependency versions via
dependabot, and our CI is using pinned dependencies. As such, I believe we
can afford to leave out the upper limit.

Other changes:
- Ignore a new warning generated by sphinx, indicating that its caching does
  not work as intended. This can be removed once https://github.com/sphinx-gallery/sphinx-gallery/issues/1286
  is resolved.

---------

Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>