Force newer version of ansys-mapdl-core
The ansys-mapdl-core package is incompatible with numpy 2.0. However,
older versions did not declare this incompatibility. To avoid
explicitly limiting the numpy version here (and thus missing when
the incompatibility is resolved), we force a sufficiently new
version of ansys-mapdl-core that does declare the incompatibility.

Otherwise, 'poetry' resolves the new numpy version, breaking the
installation of ansys-mapdl-core.