Fix EdgePropertyList construction via _from_object_info (#562)
Fixes #561

The `EdgePropertyList` retains the _same_ list instance during its lifetime 
for containing the edge property wrappers. This is done so that references
user code might hold to its items to correctly update its contents.

In contrast to other types, it means that the state is not re-fetched from
the `_pb_object` if it were to change outside of the `EdgePropertyList`'s
control.
This led to a bug when constructing objects via `_from_object_info`, which
first default-constructs the object (in the process creating the `EdgePropertyList`)
and then modifies the `_pb_object`.

To fix this, the `EdgePropertyList` now stores whether its parent object was
stored when this list was last accessed. If it changes from unstored to stored,
the following logic is applied:
- If the current list is already populated (the "regular" case), only a sanity check
  for matching length is performed
- If the current list is empty, it is re-fetched from the parent object. This is the
  case which occurs during construction with `_from_object_info`

Since the edge property list can be in an inconsistent state (empty when it shouldn't 
be) while the parent is unstored, we disallow accessing it in this state. It is still
allowed however to fully replace the contents.

This PR also fixes a bug in `tree_object_from_resource_path`, where the `channel`
argument was still used instead of `server_wrapper`.