Add step for testing on the released server (#573)
Add a step to the tests on Python 3.12 for testing with the 2024R2 server version.

Expected test failures (new features) are marked with the new `xfail_before` fixture. The
only current expected failure is the modeling ply geometry export test (added in 25.1).

Also fixes an issue with the `clone` method when `unlink=True` is specified: Any 
fields unknown to the current client were retained. This caused an error on storing,
since those may be unknown linked objects. In the current case, the newly added 
links to the `Fabric` caused this.
Calling `DiscardUnknownFields()` on the protobuf message fixes this.