Implement recursive copying helper (#563)
Implement a `recursive_copy` helper function, which:
- takes a list of source objects
- walks its children and (optionally) linked objects
- copies the sub-tree to a new location, specified by
  a dict mapping old parent to new parent [1]

Adds a dependency on `networkx`, for computing the
order in which objects should be stored s.t. all their 
dependencies (parent, linked objects) are already stored.
We may reuse the dependency graph produced for this
task in other contexts.

Note that the copy operation may fail if a linked object is present
in the API layer, but not supported by PyACP yet. In this case,
the linked object is still detected (since this is done by iterating through
the protobuf object), but the object cannot be instantiated.
I think this is acceptable, as it's a temporary problem and may be more
desirable than silently losing the link.

Other changes: 
- added a `.clone()` method to the edge property lists. This was
  used in the initial implementation of this features. Since it's 
  consistent with the API of other objects, the method is kept.
  Each edge property list type separately implements `.clone()`. [2]
- add a `.parent` property to tree objects, which instantiates / gets
  the parent via its resource path
- make tree objects hashable
- upload the XML coverage report separately for all unittest runs,
  and distinguish them (by Python version and server version) using
  flags.

[1] more than one new parent may be needed, for example
when copying a Modeling Ply (parent: Modeling Group) which
links to a Fabric (parent: Model).
[2] in general, the `GenericEdgePropertyType` classes have a
lot of code duplication; this can be dealt with in a later PR (not 
urgent).