Add type information to create_ and add_ method documentation
Provide type information for the parameters of the ``create_*``
and ``add_*`` methods in the documentation.

The __signature__ of the methods is constructed from from the wrapped
class's signature, only adding a 'self' parameter.

The 'sphinx-autodoc-typehints' extension previously used does not
support showing this dynamically created signature, so it was replaced
by using the built-in type hint support of 'sphinx.ext.autodoc'.

Return types are provided in the signature, but Sphinx incorrectly
shows the generic type (e.g. CreatableTreeObject) instead of the
one present in __signature__. Since this is not useful, the return
type is hidden unless specifically documented.

Closes #612