Enable changing the unit system (#567)
Allow setting the `unit_system` attribute on the model.

The `UnitSystemType` enum is extended by a value `FROM_FILE = "from_file"`,
which is an alias to `UNDEFINED`. This allows for using the more descriptive
`"from_file"` when loading an FE model.
When getting the `unit_system` attribute from the model, the `FROM_FILE`
is never returned; the "primary" value `UNDEFINED` is used instead, as before.