Fix teardown of direct launcher when start fails
The direct launcher '.stop()' errored when '.start()' had failed,
since the '_process' attribute was not yet defined.
This caused the error message to be misleading, as it was pointing
to the error in '.stop()' first, instead of the actual error in
'.start()'.

This is fixed by explicitly setting '_process' to None in the
constructor of the DirectLauncher class, and checking if it is
not None before trying to stop it.