Convert path to positional argument
Convert the `path` from a keword-only argument to a positional
argument in the import and export methods.

This is motivated mostly by the `import_model` method, where
the path is the only required argument. As such, it's intuitive
to call the function with the path as positional argument.

The other export methods (solid model export and solid model
skin export) are adapted for consistency.
For the model export methods, the `path` was already a positional
argument, so no changes were needed.