Add vulnerability check CI job (#650)
Add usage of the `check-vulnerabilities` action, to check for code vulnerabilities.

Checks for the use of `assert` are globally disabled. The reason for this check is that
`assert` can be misused to check for invalid input, but we use it to validate programmer
assumptions (i.e., they should never raise unless there is a bug).

For the identified issues which were ignored (related to launching the ACP gRPC
server executable), a "security considerations" page is added to the documentation.

Since we are not certain if the `check-vulnerabilities` action works on `main` with
current permissions, CI is configured to continue when it fails.