Show which examples use an object in the API documentation (#683)
Use the 'minigallery' feature of `sphinx-gallery` to show which examples use a specific feature in its API documentation.

Note that examples which are not rendered (i.e., the PyMechanical ones) do not show up. Also, uses within a function (defined in the example) do not seem to be captured.

However, where present, I think it's a nice way of linking the API reference to a usage example. When there are no examples, the section is not shown.