Split examples into groups (#684)
Split the examples into different groups for modeling features, complete workflows, and use cases.

Each group is a direct subdirectory of `doc/source/examples`. The consequence of this is that the directory name needs to be repeated:
- in the `conf.py`
- in the `make.bat` and `Makefile` for the `clean` target
 
This could be avoided by introducing an additional nesting level, at the cost of making the resulting URLs uglier.

The individual example filenames are changed to be more descriptive, addressing #527.