Make PyVista optional, fix optional deps definition (#685)
Make the `pyvista` dependency optional, adding the `plotting` extra dependency group to install it.

Add a ``requires_pyvista`` decorator, which is added to functions which import PyVista internally. The user is advised to run ``pip install ansys-acp-core[plotting]`` to install PyVista if it is missing.

Mark tests which require PyVista with ``pytest.mark.plotting``, and add a separate set of CI tests which:
- checks that PyACP can be imported if _only_ the `main` dependency group is installed
- runs the tests with only the `main` and `test` dependency groups, with `-m "not plotting"` to exclude plot tests

Remove the duplication of dependencies between the `extras` and the `dev` dependency, since that didn't work properly. Instead add ``--all-extras`` or ``--extras <extra_name>`` to ``poetry install`` where needed.

Move some internal utilities from the top-level ``ansys.acp.core`` to the ``ansys.acp.core._utils`` module.