Doc: add link to the Ansys Help Composites page
Add a link to the main Composites page in the Ansys Help as
a grid item card on the main page of the documentation.

The reason for linking this page instead of a more specific page
is that it is the only one where I could easily figure out how
to link to a page that does not change URL with every release.