Enable sensor by solid model (#788)
Make solid model and imported solid model objects linkable to the sensor, which
enables the 'sensor by solid model' feature.

Add a sensor liked to both a regular and imported solid model to the tests.

In the sensor example, add a sensor by solid model, and skip any `None` entries in
the `print_measures` helper function.

Resolves #722