Store composite definitions in UserFiles directory
In the PyMechanical integration, store the composite definitions
H5 file in the UserFiles directory of the Mechanical project
tree.
This matches the new behavior of Mechanical standalone (since 25R2),
which computes the path of this file relative to the UserFiles
directory (i.e., expects it to be inside UserFiles).

With the combined changes, projects created with the PyACP /
PyMechanical workflow can now be post-processed also within
Mechanical directly.