Add a feature tree (#821)
Add helper functions to either print a tree of ACP objects, or
print it. In the documentation, show the tree structure, and
explain how it affects the access and creation of objects.

To implement the feature tree, record metadata when creating
mapping properties (value type, and read-only status), and use
a separate _exposed_grpc_property subclass to indicate it is
a mapping.