Model / feature tree printing: set 'show_lines' default to False
Change the 'show_lines' parameter default to False in the model
and feature tree printing functions. The reason for this is
that some of the line characters are not representable in
Windows-1252 encoding, which will cause an exception when this
encoding is used for printing.

Adapt the documentation to instead pass the 'show_lines=True'
parameter explicitly.

Make the 'to_string' method of the Node class public, to allow
passing the 'show_lines' parameter to it. Otherwise, the tree
without lines could only be printed, not obtained as a string.