Re-enable the 'price' attribute on ModelElementalData (#853)
For server version 25.2 or later, re-enable the 'price' attribute
on ModelElementalData.
In order to know the server version, the _field_names method now
takes an optional server_version argument. In the 'ModelElementalData',
the 'price' field is removed if the server version is less than 25.2.

Add back the previous test for the 'price' attribute, and a model
which crashed when accessing the 'price' with 2025R1 and earlier.

Closes #720.