Add how-to for copying objects (#854)
Add a how-to guide which explains how to use the recursive_copy
function. The guide builds up understanding of the function
parameters by showing examples of their use.

Add a parameter ``label_by_id`` to the model printer, since it
is more convenient to know the ID rather than the name when
trying to access objects in the model. Additionally, it allows
distinguishing between the original object and its copy.

Modify the doctest flags to normalize whitespace. Without this,
the tree print output failed to match.

Closes #830.