Allow stackup thickness and area weight to be None (#873)
When the fabrics in a stackup do have a material, or the stackup
has no fabrics, the 'thickness' and 'area_weight' properties
cannot be computed. This change allows these properties to be
'None' in these cases.
This case is only supported in the 26.1 server version or later.
Earlier versions would raise an error when getting the properties
under these conditions.