Convert DAT file to CDB for import into Mechanical (#896)
Context:

In the 'CDB to Mechanical' workflow, using a DAT file (as exported
from Mechanical) can fail when e.g. the layup only covers part of
the mesh (see #826).
This can be fixed by converting the DAT file to a CDB file
using MAPDL. This improves the reliability of loading the model
into Mechanical, and allows re-enabling the 'valid blocked CDB'
check in the import.

Changes:

- Enable the 'valid blocked CDB' check in the `import_acp_mesh_from_cdb`
  helper function.
- Update the 'Create input file' documentation to include the conversion
  from DAT to CDB using (Py)MAPDL.
- Add a conversion from DAT to CDB to the 'CDB to Mechanical' workflow
  example.

Closes #826.

Co-authored-by: René Roos <105842014+roosre@users.noreply.github.com>