Implement gRPC transport options (#918)
For each launch mode, implement the gRPC transport options which are applicable:
- `direct`, `connect` and `connect_local`: mTLS, UDS (Unix only), WNUA (Windows only), and insecure
- `docker_compose`: mTLS and insecure

The corresponding options are added to the launch configuration dataclasses, and then
converted to the (mode-specific) transport options classes provided by `ansys-tools-local-product-launcher`.

When the server is unpatched, only the `insecure` transport mode is accepted, the others
raise an error (either when launching the server, or when attaching). One exception is
WNUA in `connect` or `connect_local` mode, since there's no way for the client to distinguish
it from `insecure` mode (the check is purely server-side).

With a patched server, the direct launch mode should work out of the box without code changes.

Add documentation for how to change the gRPC transport mode when launching ACP.